Fix compiler warnings and apply code formatting
- Fix sign comparison warnings in log_site.hpp and main.cpp by adding appropriate casts
- Suppress -Winterference-size warning for hardware_destructive_interference_size with GCC pragmas
- Fix unused variable warning in simple.cpp by conditionally compiling assignment
- Apply consistent code formatting across all header files
- Clean up unnecessary code in log_writers.hpp (simplified get_next_fd() calls)
- Remove unused variable assignments and improve type safety

+++ include/log_site.hpp
@@ -74,7 +74,10 @@ struct log_site_registry
         std::lock_guard<std::mutex> lock(registry_mutex());
         sites().emplace_back(log_site_descriptor{file, line, function ? function : "unknown", min_level});
 
-        if (file && std::strlen(file) > longest_file()) { longest_file() = static_cast<int32_t>(std::strlen(file)); }
+        if (file && static_cast<int32_t>(std::strlen(file)) > longest_file())
+        {
+            longest_file() = static_cast<int32_t>(std::strlen(file));
+        }
         return sites().back();
     }
 
@@ -216,13 +219,13 @@ struct log_site_registry
      *          with thousands of LOG() statements, this can be expensive.
      *          Each LOG() statement that survives compile-time filtering creates one entry.
      * @note Thread-safe: returns a snapshot at the time of the call
-     * 
+     *
      * Example:
      * @code
      * auto all_sites = log_site_registry::get_all_sites();
      * std::cout << "Total sites: " << all_sites.size() << "\n";
      * for (const auto& site : all_sites) {
-     *     std::cout << site.file << ":" << site.line 
+     *     std::cout << site.file << ":" << site.line
      *               << " level=" << log_level_names[static_cast<int>(site.min_level)]
      *               << " function=" << site.function << "\n";
      * }
